commit laptop

+++ NCKH-QLDA/src/NCKH-QLDA-Core/WebSite.Core.Domain/ModelMeta/ChiTietDeTaiMeta.cs
@@ -6,7 +6,7 @@ namespace WebSite.Core.Domain.ModelMeta
 {
     public class ChiTietDeTaiMeta
     {
-        public double? DiemSo { get; set; }
+        public float? DiemSo { get; set; }
         public string NhanXet { get; set; }
     }
 }
